🚚 Rename: itemDetail로 폴더명 변경
